Remove exec flag from library-style script modules.
They might confuse users which parts are meant to be executed from the command
line, plus it is not required as they are "sourced" in in the main "setup.sh"
script. Thus it is better not to have them executable anyway.

If required, individual modules can still be called using "source
<modulename>.sh" if all required variables are set appropriately.

refs #97