Fix function to work with multiple packages and newer dpkg versions.
Grepping for the error message was not very portable, i.e. it did not work with
dpkg in Ubuntu 13.10 any more as the error message slightly changed. Also the
formatting is different, so the cut statement was also broken. Additionally,
the old function only worked properly with single packages. Thus, checking for
the packages individually and testing for the "installed" status should be a
lot more robust.

refs #97